archive/2019_12_02_cleaning/requirements.txt
appnope==0.1.0
backcall==0.1.0
biopython==1.73
bleach==3.1.0
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.3
et-xmlfile==1.0.1
imageio==2.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kiwisolver==1.0.1
lxml==4.3.3
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.16.0
openpyxl==2.6.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==17.1.2
qtconsole==4.4.3
rfpimp==1.3.3
scikit-image==0.15.0
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
shap==0.28.5
six==1.12.0
sklearn==0.0
stevedore==1.30.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.29.1
traitlets==4.3.2
virtualenv==16.3.0
virtualenv-clone==0.5.1
virtualenvwrapper==4.8.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xlrd==1.2.0
